rename package ts
