+++ src/com/alds/quiz/tsf/Tag.java
@@ -1,4 +1,4 @@
-package com.alds.quiz.ts;
+package com.alds.quiz.tsf;
 
 class Tag {
 	/**
