+++ src/com/alds/quiz/tsf/TagSizeTestCase.java
@@ -1,4 +1,4 @@
-package com.alds.quiz.ts;
+package com.alds.quiz.tsf;
 
 import java.util.List;
